requirements.txt
fastapi[standard]
pydantic
websockets
python-jose[cryptography] 
passlib[bcrypt]
python-multipart
psycopg2-binary
python-dotenv